move common components to UI folder
