+++ src/Landing/Experiences/Card.js
@@ -1,7 +1,7 @@
 import React from "react";
 import styled from "styled-components";
 
-import { CardLink, CardImg } from "../../styled";
+import { CardLink, CardImg } from "../../UI/styled";
 import Stars from "../../Stars";
 
 const DetailsRow = styled.div`
