+++ src/Landing/Explore/Card.js
@@ -1,7 +1,7 @@
 import React from "react";
 import styled from "styled-components";
 
-import { CardLinkRouter, CardImg } from "../../styled";
+import { CardLinkRouter, CardImg } from "../../UI/styled";
 
 const Card = styled(CardLinkRouter)`
   border: 1px solid rgba(72, 72, 72, 0.2);
